apps/api/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --all-extras -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   mcp
    #   openai
    #   sse-starlette
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
azure-core==1.35.0
    # via azure-identity
azure-identity==1.24.0
    # via ai-assist-api (pyproject.toml)
basedpyright==1.31.3
    # via ai-assist-api (pyproject.toml)
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via uvicorn
colorama==0.4.6
    # via griffe
coverage==7.10.6
    # via pytest-cov
cryptography==45.0.7
    # via
    #   azure-identity
    #   msal
    #   pyjwt
distro==1.9.0
    # via openai
fastapi==0.116.1
    # via ai-assist-api (pyproject.toml)
griffe==1.13.0
    # via openai-agents
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   ai-assist-api (pyproject.toml)
    #   langsmith
    #   mcp
    #   openai
httpx-sse==0.4.1
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
jiter==0.10.0
    # via openai
jsonschema==4.25.1
    # via mcp
jsonschema-specifications==2025.4.1
    # via jsonschema
langsmith==0.4.21
    # via ai-assist-api (pyproject.toml)
mcp==1.13.1
    # via openai-agents
msal==1.33.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
nodejs-wheel-binaries==22.18.0
    # via basedpyright
openai==1.102.0
    # via openai-agents
openai-agents==0.2.10
    # via
    #   ai-assist-api (pyproject.toml)
    #   langsmith
orjson==3.11.3
    # via langsmith
packaging==25.0
    # via
    #   langsmith
    #   pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   ai-assist-api (pyproject.toml)
    #   fastapi
    #   langsmith
    #   mcp
    #   openai
    #   openai-agents
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via
    #   ai-assist-api (pyproject.toml)
    #   mcp
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via msal
pytest==8.4.1
    # via
    #   ai-assist-api (pyproject.toml)
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==1.1.0
    # via ai-assist-api (pyproject.toml)
pytest-cov==6.2.1
    # via ai-assist-api (pyproject.toml)
python-dotenv==1.1.1
    # via
    #   ai-assist-api (pyproject.toml)
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via mcp
pyyaml==6.0.2
    # via uvicorn
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   azure-core
    #   langsmith
    #   msal
    #   openai-agents
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
ruff==0.12.11
    # via ai-assist-api (pyproject.toml)
six==1.17.0
    # via azure-core
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sse-starlette==3.0.2
    # via mcp
starlette==0.47.3
    # via
    #   fastapi
    #   mcp
tqdm==4.67.1
    # via openai
types-requests==2.32.4.20250809
    # via openai-agents
typing-extensions==4.15.0
    # via
    #   azure-core
    #   azure-identity
    #   fastapi
    #   openai
    #   openai-agents
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via
    #   requests
    #   types-requests
uvicorn==0.35.0
    # via
    #   ai-assist-api (pyproject.toml)
    #   mcp
# uvloop==0.21.0  # Not available on Windows
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
zstandard==0.24.0
    # via langsmith
